Refatoração e correções nas funções relacionadas a armazenamento e formatação de mensagens.
